Dissolve the deformable grab-bag module into per-solver and per-capability modules.
