+++ tests/deformable/test_mpm.py
@@ -1,0 +1,131 @@
+import pytest
+import torch
+
+import genesis as gs
+
+
+@pytest.mark.required
+def test_particle_constraints(show_viewer):
+    scene = gs.Scene(
+        sim_options=gs.options.SimOptions(
+            dt=2e-3,
+            substeps=20,
+        ),
+        mpm_options=gs.options.MPMOptions(
+            lower_bound=(-1.0, -1.0, 0.0),
+            upper_bound=(1.0, 1.0, 1.0),
+            grid_density=64,
+        ),
+        show_viewer=show_viewer,
+        show_FPS=False,
+    )
+    scene.add_entity(gs.morphs.Plane())
+    rigid_box = scene.add_entity(
+        gs.morphs.Box(
+            pos=(0, 0, 0.55),
+            size=(0.12, 0.12, 0.05),
+            fixed=True,
+        ),
+    )
+    mpm_cube = scene.add_entity(
+        material=gs.materials.MPM.Elastic(
+            E=5e4,
+            nu=0.3,
+            rho=1000,
+        ),
+        morph=gs.morphs.Box(
+            pos=(0, 0, 0.35),
+            size=(0.15, 0.15, 0.15),
+        ),
+    )
+    scene.build(n_envs=2)
+
+    # Test get_particles_in_bbox - returns (n_envs, n_particles) mask
+    mask = mpm_cube.get_particles_in_bbox((-0.08, -0.08, 0.41), (0.08, 0.08, 0.44))
+    assert mask.shape == (2, mpm_cube.n_particles), "mask should be (n_envs, n_particles)"
+    assert mask.any(), "bbox should select some particles"
+    assert not mask.all(), "bbox should not select all particles"
+
+    # Attach and test following
+    link_idx = rigid_box.links[0].idx
+    mpm_cube.set_particle_constraints(mask, link_idx, stiffness=1e5)
+    initial_rigid_pos = rigid_box.get_pos()
+    initial_mpm_x = mpm_cube.get_particles_pos()[:, mask[0], 0].mean()
+
+    pos_diff = torch.tensor([0.2, 0, 0], device=gs.device)
+    rigid_box.set_pos(initial_rigid_pos + pos_diff, zero_velocity=False)
+    for _ in range(30):
+        scene.step()
+
+    mpm_diff = mpm_cube.get_particles_pos()[:, mask[0], 0].mean() - initial_mpm_x
+    assert mpm_diff > pos_diff[0] * 0.3, f"MPM should follow rigid link. Got {mpm_diff:.3f}"
+
+
+@pytest.mark.required
+@pytest.mark.parametrize("variant", ["svd", "no_svd"])
+def test_perf_dispatch(variant, show_viewer):
+    if variant == "svd":
+        elastic_model = "corotation"
+        liquid_viscous = True
+    else:
+        elastic_model = "neohooken"
+        liquid_viscous = False
+
+    scene = gs.Scene(
+        sim_options=gs.options.SimOptions(
+            dt=2e-3,
+            substeps=10,
+        ),
+        mpm_options=gs.options.MPMOptions(
+            lower_bound=(-0.4, -0.4, -0.05),
+            upper_bound=(0.4, 0.4, 0.5),
+            grid_density=64,
+        ),
+        show_viewer=show_viewer,
+    )
+    scene.add_entity(
+        morph=gs.morphs.Plane(),
+    )
+    elastic = scene.add_entity(
+        morph=gs.morphs.Box(
+            pos=(-0.1, 0.0, 0.2),
+            size=(0.08, 0.08, 0.08),
+        ),
+        material=gs.materials.MPM.Elastic(
+            E=3e5,
+            nu=0.2,
+            rho=200.0,
+            model=elastic_model,
+        ),
+    )
+    liquid = scene.add_entity(
+        morph=gs.morphs.Box(
+            pos=(0.15, 0.0, 0.2),
+            size=(0.08, 0.08, 0.08),
+        ),
+        material=gs.materials.MPM.Liquid(
+            viscous=liquid_viscous,
+        ),
+    )
+    scene.build(n_envs=2)
+
+    # Aggregate SVD flag must match the variant for the dispatch path under test to actually run.
+    assert scene.sim.mpm_solver.needs_svd == (variant == "svd")
+
+    init_elastic_pos = elastic.get_particles_pos()
+    init_liquid_pos = liquid.get_particles_pos()
+
+    for _ in range(100):
+        scene.step()
+
+    final_elastic_pos = elastic.get_particles_pos()
+    final_liquid_pos = liquid.get_particles_pos()
+
+    # Gravity acted: particles moved down on average.
+    assert init_elastic_pos[..., 2].mean() - final_elastic_pos[..., 2].mean() > 0.05
+    assert init_liquid_pos[..., 2].mean() - final_liquid_pos[..., 2].mean() > 0.05
+
+    # No ground penetration (plane is at z=0). Non-viscous liquid spreads thin enough that a sub-grid-cell penetration
+    # is normal for the MPM coupling, so allow up to 2mm.
+    assert final_elastic_pos[..., 2].min() > -1e-3
+    assert final_liquid_pos[..., 2].min() > -2e-3
